Fix for bug 242.
We warn if the user misplaces a binding annotation. In Scala, binding annotations are applied to accessor methods automatically, and there's nothing the user can do (to suppress the warning). So we special case it and don't warn. There might be a tiny number of false negatives; this shouldn't really be a problem.

git-svn-id: https://google-guice.googlecode.com/svn/trunk@707 d779f126-a31b-0410-b53b-1d3aecad763e