Remove (most) accumulation of multiple errors during provisioning
We will still collect multiple errors during configuration, but during
provisioning there are fewer benefits and in highly connected binding graphs
this can result in an error being reported once for every path to a broken
binding.  Even with this change multiple errors can be collected while injecting multiple members of a single object and in a few other less common ways (returning null from a non-nullable injection point).  But this cl removes the most common way.

This is also an enabling change.  With this it should be possible to move all
calls to Errors.withSource into catch blocks in the factory objects. So that
instead of recording progress as we provision, we will only record the sources
when the provision actually fails.  This will save an allocation for every node
and edge in the binding graph. (this change will happen in a followup)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=173327556