Calculate dependency information for errors as errors propagate instead of eagerly as provisions are executed.
The goal is to move the cost of tracking these data structures out of the fast path on the assumption that provisioning rarely fails and so the cost of calling Errors.withSource (and Errors.size()/Errors.throwIfNewErrors()....) is mostly wasted.  Instead we can add catch blocks that do this work.

This has 2 immediate consequences:
1. since the provisioning stack is calculated as exceptions propagate it will be incomplete if the exception is caught 'early'.  For example, previously if a ProvisionListener caught a ProvisionException, the exception would show, via the chain of sources, the full path to the failed provision. Now it will only show the path 'below' where it was caught.  updated tests in ProvisionListenerTest demonstrate this
2. A number of places where multiple errors used to be collected now no longer do (for example, the first member injector to throw will short circuit other members injectors).  This is an extension of a prior CL which removed multiple error collection from SingleParameterInjector.

This builds on the InternalProvisionException introduced in an earlier patch.  The change is fairly mechanical, but since this changes the calling convention for Guice, all points which participate need to be updated together.  In particular,

* the signature of InternalFactory.get(...) was changed by removing the Errors object and changing the 'throws' signature.
* All Errors.withSource(...) calls were changed to InternalProvisionException.addSource calls in a catch block.

Benchmarks show that this drops provisioning time by about 20% and allocated objects by about the same (no more Errors objects!).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178414361