Delete TODO to make `@Inject` required to use assisted inject.
Since it's actually possible to use assisted inject with field or method injection, an `@Inject` annotated constructor is not necessary.

This does mean that assisted injection does not respect the `binder().requireAtInjectOnConstructors()` requirement and it is non-trivial to propagate the injector option down to enforce this because the injection point is looked up at module configuration time.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=326730924